+++ src/main/java/interfaceadapters/CourseEvaluatorInterfaceAdapters/EvaluatorController.java
@@ -1,7 +1,7 @@
-package interfaceadapters.CourseEvaluator;
+package interfaceadapters.CourseEvaluatorInterfaceAdapters;
 
-import usecases.CourseEvaluator.EvaluatorInputBoundary;
-import usecases.CourseEvaluator.EvaluatorRequestModel;
+import usecases.CourseEvaluatorUseCase.EvaluatorInputBoundary;
+import usecases.CourseEvaluatorUseCase.EvaluatorRequestModel;
 
 public class EvaluatorController {
         private EvaluatorInputBoundary interactor;
